Create Integrations and add link unfurling
Implement content fetching

Originally implemented a SAX parser to avoid the body size limit,
but it turns out that Nokogiri doesn't support SAX parsers for HTML5.
This was quite confusing because the parser is called HTML::SAX::Parser
and everywhere else HTML is aliased to HTML5.
I had to abandon this approach after spending a few hours on it, and
then opted for the same approach that Campfire uses - fetch the whole
body and parse it. But I lowered the body size to 2MB after looking
up what the median web page body size is - 160KB 50%%, 285KB 90%%.